Code review

+++ src/subscriptions/subscriptions.service.ts
@@ -1,19 +1,15 @@
-import {
-  Injectable,
-  InternalServerErrorException,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { DataSource, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { SubscriptionEntity } from './entities/subscription.entity';
 import { HttpException, HttpStatus } from '@nestjs/common';
+import { UserEntity } from 'src/user/entities/user.entity';
 
 @Injectable()
 export class SubscriptionsService {
   constructor(
     @InjectRepository(SubscriptionEntity)
     private readonly repository: Repository<SubscriptionEntity>,
-    private dataSource: DataSource,
   ) {}
 
   async create(id: number, followingId: number) {
@@ -55,8 +51,9 @@ export class SubscriptionsService {
   }
 
   async findPopularUsers(skip = 0, take = 10) {
-    const result = await this.repository.query(
-      `
+    const result: UserEntity & { subscribers: number }[] =
+      await this.repository.query(
+        `
       SELECT users.*, COUNT("followingId") as subscribers
       FROM subscriptions
       JOIN users ON users.id = subscriptions."followerId"
@@ -64,8 +61,8 @@ export class SubscriptionsService {
       ORDER BY subscribers DESC
       OFFSET $1 LIMIT $2
     `,
-      [skip, take],
-    );
+        [skip, take],
+      );
 
     return result.map((user) => {
       delete user['password'];
@@ -74,31 +71,18 @@ export class SubscriptionsService {
   }
 
   async remove(subscriptionId: number) {
-    const queryRunner = this.dataSource.createQueryRunner();
-    await queryRunner.connect();
-    await queryRunner.startTransaction();
+    const subscription = await this.repository.findOneBy({
+      id: subscriptionId,
+    });
 
-    try {
-      await this.repository.delete(subscriptionId);
-
-      await queryRunner.commitTransaction();
-
-      return {
-        message: `Запись подписки с id ${subscriptionId} успешно удалена`,
-      };
-    } catch (error) {
-      await queryRunner.rollbackTransaction();
-      handleMethodErrors(error, subscriptionId);
-    } finally {
-      await queryRunner.release();
+    if (!subscription) {
+      throw new NotFoundException('Подписка не найдена');
     }
-  }
-}
 
-function handleMethodErrors(error: any, id: number) {
-  if (error.status == 404) {
-    throw new NotFoundException(`Пользователь с id ${id} не найден`);
-  } else {
-    throw new InternalServerErrorException('Произошла серверная ошибка');
+    await this.repository.delete(subscriptionId);
+
+    return {
+      message: `Запись подписки с id ${subscriptionId} успешно удалена`,
+    };
   }
 }
